making the tools sepreately
